geek/tos3-migration: add fixes for not executed post-install hooks
Updater is not running post-install hooks in appropriate environemnt.
That causes host of the post-install hooks to be skipped. These comments
cover known issues users can encounter because of this.

The version of updater included in Turris OS 5.2.0 should fix this so
with release of TOS 5.2.0 to HBS we can revert this.