reforis: Add page about DNS settings (including custom forwarders)
It includes a part of the page into the original page about DNS in Foris.
The new page is based on the original one for Foris. A part of it is
directly moved from it and then included back (using the include-markdown
plugin). It means that this (Foris/reForis-independent) part can be
maintained only at one place.