basics/changelog: Do not remove last line from changelog message
The changelog message taken from the tags in the turris-build repository
are signed often and so the message ends on:
```
-----END PGP SIGNATURE-----
```

However, if the tag is not signed, the last line is the last line of the
changelog so this `pop()` will remove the important data.

I don't see why to pop the last element at all, as the format of the
line will not pass the regex.

On top of it, remove console log messages which are not needed in
production.