treewide: Turris OS repositories on our GitLab were moved
Before this change as repositories were changed, API does not count with
that, which results in HTML error 404 - Not found while showing
changelog, so we need to fix it here as well.

Also, it fixes this warning on repositories turris/os/build and
turris/os/packages:

Project 'turris/turris-os-packages' was moved to 'turris/os/packages'.
Please update any links and bookmarks that may still have the old path.