geek: increase min. storage of microSD on Turris 1.x's Btrfs migration
We decided that we will increase necessary free storage on microSD which
users are using on Turris 1.x from 4 GB to 8 GB, because of these
reasons:

- Turris Omnia has 8 GB internal storage
- Turris MOX, by default, has inserted a 16 GB microSD card

These days microSD cards are relatively cheap to buy in decent capacity with
good I/O with the Application Performance Class. It is better to invest in a
higher capacity than deal with full storage.

Also, in the past, we gave users to help them with migration to Btrfs 8 GB microSD card
in some local events.