btrfs_turris1x: switch from migration to factory reset
The same steps have to be performed for factory reset and in general it
is more important to explain users how to perform factory reset for
Turris 1.x routers on SD card than how to migrate now.

This also fixes problem that kernel would not match modules. The kernel
has to be updated from factory before rollback.